refactor: ChessGame의 디렉토리 위치를 domain으로 변경
